requirements.txt
openai>=1.12.0
tqdm
Pillow>=8.0.0

# Additional ML dependencies
numpy>=1.21.0
safetensors>=0.3.0
tokenizers>=0.15.0
einops>=0.6.0
sentence-transformers>=2.2.0
scikit-learn>=1.0.0

# PyTorch dependencies - required for sentence-transformers
torch>=2.0.0,<2.5.0
torchvision>=0.15.0
transformers>=4.37.0

# Optional but recommended for better performance
# Qwen 2.5 VL model dependencies
# accelerate>=0.26.0
qwen-vl-utils>=0.0.3
# flash-attn>=2.0.0;sys_platform!="darwin"